[fixed] Power is withdrawn from inventory when depositing to harvester.
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@2746 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/managers/resource/ResourceManager.h
@@ -166,9 +166,9 @@ public:
 
 	void harvestResourceToPlayer(PlayerCreature* player, ResourceSpawn* resourceSpawn, const int quantity);
 
-	unsigned long long getAvailablePowerFromPlayer(PlayerCreature* player);
+	unsigned int getAvailablePowerFromPlayer(PlayerCreature* player);
 
-	void removePowerFromPlayer(PlayerCreature* player, unsigned long long power);
+	void removePowerFromPlayer(PlayerCreature* player, unsigned int power);
 
 	void getResourceListByType(Vector<ManagedReference<ResourceSpawn* > >& list, int type, int zoneid);
 
@@ -242,9 +242,9 @@ public:
 
 	void harvestResourceToPlayer(PlayerCreature* player, ResourceSpawn* resourceSpawn, const int quantity);
 
-	unsigned long long getAvailablePowerFromPlayer(PlayerCreature* player);
+	unsigned int getAvailablePowerFromPlayer(PlayerCreature* player);
 
-	void removePowerFromPlayer(PlayerCreature* player, unsigned long long power);
+	void removePowerFromPlayer(PlayerCreature* player, unsigned int power);
 
 	void getResourceListByType(Vector<ManagedReference<ResourceSpawn* > >& list, int type, int zoneid);
 
@@ -325,9 +325,9 @@ public:
 
 	void harvestResourceToPlayer(PlayerCreature* player, ResourceSpawn* resourceSpawn, const int quantity);
 
-	unsigned long long getAvailablePowerFromPlayer(PlayerCreature* player);
+	unsigned int getAvailablePowerFromPlayer(PlayerCreature* player);
 
-	void removePowerFromPlayer(PlayerCreature* player, unsigned long long power);
+	void removePowerFromPlayer(PlayerCreature* player, unsigned int power);
 
 	void createResourceSpawn(PlayerCreature* playerCreature, const String& restype);
